# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Notify Module",
        "contact": {
            "email": "pvalente@onesource.pt"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/emailTemplate/{email_template_key}": {
            "get": {
                "tags": [
                    "Email Template"
                ],
                "summary": "Show a Email Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email_template_key",
                        "in": "path",
                        "description": "Email Template Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "LANG-CODE",
                        "in": "header",
                        "description": "User Language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "LANG-CODE-DEFAULT",
                        "in": "header",
                        "description": "Entity default Language",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show the Email Template data",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Template not Found",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Email Template"
                ],
                "summary": "Update an Email Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Body",
                        "in": "body",
                        "description": "Email Template Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emailTemplateCreate"
                        }
                    },
                    {
                        "name": "email_template_key",
                        "in": "path",
                        "description": "Email Template Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The updated Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailGroupReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Template not Found",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to update Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Email Template"
                ],
                "summary": "Delete an Email Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "email_template_key",
                        "in": "path",
                        "description": "Email Template Key",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateDeleteReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Template not Found",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to delete Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    }
                }
            }
        },
        "/emailTemplate": {
            "post": {
                "tags": [
                    "Email Template"
                ],
                "summary": "Create an Email Template",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Body",
                        "in": "body",
                        "description": "Email Template Data",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emailTemplateCreate"
                        }
                    },
                    {
                        "name": "X-AUTH-TOKEN",
                        "in": "header",
                        "description": "User Auth Token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "X-MODULE-TOKEN",
                        "in": "header",
                        "description": "Module Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the newly created Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateReply"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "404": {
                        "description": "Email Template not found",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    },
                    "500": {
                        "description": "Failed to store Email Template",
                        "schema": {
                            "$ref": "#/definitions/emailTemplateErrorDefault"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emailTemplateReply": {
            "type": "object",
            "allOf": [
                {
                    "properties": {
                        "id": {
                            "type": "integer",
                            "format": "integer"
                        },
                        "email_template_key": {
                            "type": "string",
                            "format": "string"
                        },
                        "type_id": {
                            "type": "integer",
                            "format": "integer"
                        },
                        "email_group_id": {
                            "type": "integer",
                            "format": "integer"
                        },
                        "subject": {
                            "type": "string",
                            "format": "string"
                        },
                        "header": {
                            "type": "string",
                            "format": "string"
                        },
                        "content": {
                            "type": "string",
                            "format": "string"
                        },
                        "footer": {
                            "type": "string",
                            "format": "string"
                        },
                        "type": {
                            "type": "object",
                            "allOf": [
                                {
                                    "$ref": "#/definitions/typeReply"
                                }
                            ]
                        },
                        "site_key": {
                            "type": "string",
                            "format": "string"
                        },
                        "created_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "updated_at": {
                            "type": "string",
                            "format": "date"
                        },
                        "deleted_at": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                }
            ]
        },
...
    },
...
}